requirements.txt
python-telegram-bot
python-dotenv
apscheduler
fastapi
uvicorn
